2.1.0 - Read Changelog

+++ recipes/kitchen_shutdown.rb
@@ -6,6 +6,8 @@ when 'linux'
     hour node['bonusbits_base']['kitchen_shutdown']['hour'].to_s
     user 'root'
     command 'shutdown -h now >/dev/null 2>&1'
+    not_if { node['bonusbits_base']['deployment_type'] == 'docker' }
+    only_if { node['bonusbits_base']['deployment_method'] == 'kitchen' }
   end
 else
   return
